requirements.txt
# Few-Shot Sketch Recognition Framework Requirements

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0

# Data Processing
numpy>=1.24.0
pillow>=9.5.0
opencv-python>=4.7.0
scikit-image>=0.20.0

# Machine Learning Utilities
scikit-learn>=1.2.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0

# Progress Bars and Utilities
tqdm>=4.65.0
pyyaml>=6.0

# Data Download (for QuickDraw and TU-Berlin)
requests>=2.31.0
gdown>=4.7.1
kaggle>=1.5.15

# Augmentation
albumentations>=1.3.0

